Another optimization: the cumsum
